Move helpers functions to 'utils' and update code to use them
