+++ utils/helpers.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package v1alpha2
+package utils
 
 func PointerOf[A any](a A) *A {
 	return &a
